* release: (338 commits)
Update: Grunt
Update: Grunt string-replace
Update: Version name update
updated change log
Added change log
PROD-8149 - Pulled API changes
[PROD-8240] grunt
PROD-8378 - Grunt
PROD-8378 - Remove code which strip html tags so script tag can remove from draft also fixed jshint
PROD-8378 - sanitize topic title when added in draft
PROD-8378 -  sanitize topic title when add/edit
[PROD-8240] Fix moderations table UI issues in mobile screen
[PROD-8240] Fix Group and Activity tables UI issues in mobile screen
[PROD-8240] Fix error overflow outside container in smaller screens
[PROD-8332] Remove open class from message more option wrap once the dropdown is closed
[PROD-8310] Fix create new message uploader button UI issue
[PROD-8210] Set single profile avatars height to auto for better aspect ratio handling
Update: Grunt
Update: Grunt string-replace
Update: Version name update
...

+++ src/bp-media/bp-media-filters.php
@@ -831,7 +831,11 @@ function bp_media_forums_save_gif_data( $post_id ) {
 			bp_activity_update_meta( $main_activity_id, '_gif_data', $gdata );
 			bp_activity_update_meta( $main_activity_id, '_gif_raw_data', $gif_data );
 		}
-	} else {
+	} elseif (
+		isset( $_POST['action'] ) &&
+		in_array( $_POST['action'], array( 'bbp-edit-reply', 'bbp-edit-topic', 'bbp-edit-forum' ), true ) &&
+		empty( $_POST['bbp_media_gif'] )
+	) {
 		delete_post_meta( $post_id, '_gif_data' );
 		delete_post_meta( $post_id, '_gif_raw_data' );
